Move newer functions to common rpc interface
